Añadida get_est_arranque_p() y corregida get_est_arranque() get_est_arranque_p() devuelve el nombre del estado arranque entre comillas para la impresión

+++ P7/estado_t.hpp
@@ -46,7 +46,7 @@ class estado_t {
   void print_trans()const;
 
   //GETTER
-  unsigned int get_id() const { return id_; }
+  long long int get_id() const { return id_; }
   std::string get_name() const {return name_;}
   bool get_acept() const { return acept_; }
   int is_arranque() const { return arranque; }
